Client: Renamed followers-or-following modal and styles to followers-or-following-or-likes. Added property to the component to determine whether the modal is for displaying post-liking users, and should be labeled with 'Likes' instead of 'Followers' or 'Following'
