+++ client/src/pages/my-profile/my-profile-page.component.tsx
@@ -76,7 +76,7 @@ import {
 import PostTile from '../../components/post-tile/post-tile.component';
 import PostModal from '../../components/post-modal/post-modal.component';
 import PostOrCommentOptionsModal from '../../components/post-or-comment-options-modal/post-or-comment-options-modal.component';
-import FollowersOrFollowingModal from '../../components/followers-or-following-modal/followers-or-following-modal.component';
+import FollowersOrFollowingOrLikesModal from '../../components/followers-or-following-or-likes-modal/followers-or-following-or-likes-modal.component';
 
 import './profile-page.styles.scss';
 
@@ -525,7 +525,7 @@ export const MyProfilePage: React.FC<MyProfilePageProps> = ({
         }}
         isCurrentUserPostOrComment={currentUserPostOrComment}
       />
-      <FollowersOrFollowingModal
+      <FollowersOrFollowingOrLikesModal
         users={isFollowersModal ? followersArray : usersFollowingArray}
         show={followersOrFollowingModalShow}
         onHide={() => {
